address codeQL warnings (#598)
* codeql fixes

* fix type mismatches

* fix pointers in w32_time methods

* fixes for codeQL warnings

* modify checks for codeql warnings

* add comments for codeql suppressions

* additional codeql fixes and suppressions

* add codeql fixes

* add comments for codeql

* add comments for codeql

* switch from debug to error log messages

* fix another merge conflict

fix line endings in gss-sspi.c

* add null check in channels.c

* address PR feedback

* address additional review feedback

* add CodeQL comments to common code

* fix unittest-win32compat

* fix unit test

* address review feedback

* remove suppression

+++ ed25519.c
@@ -106,7 +106,7 @@ int crypto_sign_ed25519_open(
     const unsigned char *pk
     )
 {
-  unsigned int i;
+  unsigned long long i; // fix CodeQL SM01735
   int ret;
   unsigned char t2[32];
   ge25519 get1, get2;
